VNC-500 Refactor SCTP packet parsing and chunk processing with enhanced validation

+++ common/network/webudp/CRC32.h
@@ -4,4 +4,4 @@
 #include <stdint.h>
 
 uint32_t StunCRC32(const void* data, int32_t len);
-uint32_t SctpCRC32(const void* data, int32_t len);
+uint32_t SctpCRC32(const void* data, size_t len);
